+++ Pascal_Triangle_II.cpp
@@ -24,7 +24,7 @@ int main()
 {
 	Solution so;
 	vector<int> ret = so.getRow(5);
-	for(int i=0; i<ret.size();i++)
+	for(size_t i=0; i<ret.size();i++)
 	{
 		cout<<ret[i];
 	}
